refactor(memory): CRUD 移入基类 + _validate_all_items + 更新架构文档
- base.py: 新增 _write_lock 抽象原语，CRUD 方法从抽象改为基于三个原语的具体实现
- base.py: 新增 _validate_all_items() 介质无关的字段校验方法
- yaml.py: 删除重复 CRUD 代码 (~60行)，仅保留 _load_all/_save_all/_write_lock + self_check
- docs/backend-architecture/memory-layer.md: 同步更新架构描述、继承图、代码片段
