+++ api/memory/manager/base.py
@@ -3,6 +3,7 @@
 import datetime
 import secrets
 from abc import ABC, abstractmethod
+from contextlib import AbstractContextManager
 from typing import Any, TypedDict
 
 
@@ -92,7 +93,7 @@ class BaseMemoryManager(ABC):
     """记忆管理器抽象接口。
 
     所有记忆存储后端（YAML、数据库等）均应继承此类。
-    子类只需实现 _load_all() 和 _save_all() 两个原语，
+    子类只需实现 _load_all()、_save_all()、_write_lock() 三个原语，
     即可继承 add / delete / update / merge 等公共方法。
     """
 
@@ -106,6 +107,11 @@ class BaseMemoryManager(ABC):
     @abstractmethod
     def _save_all(self, items: dict[str, MemoryItem]) -> None:
         """覆写全部条目。子类须自行保证并发安全。"""
+        ...
+
+    @abstractmethod
+    def _write_lock(self) -> AbstractContextManager[None]:
+        """写操作的锁上下文。子类必须覆盖以实现并发控制。"""
         ...
 
     # ── 启动自检 ────────────────────────────────────────────
@@ -124,6 +130,43 @@ class BaseMemoryManager(ABC):
             issues（不可修复问题）、repaired（已修复问题）、item_count（有效条目数）。
         """
         ...
+
+    def _validate_all_items(
+        self, items: dict[str, MemoryItem]
+    ) -> tuple[list[str], list[str]]:
+        """校验所有条目的 MemoryItem 字段完整性，返回 (issues, repaired)。
+
+        介质无关，子类的 self_check 可直接调用此方法以复用校验逻辑。
+        """
+        issues: list[str] = []
+        repaired: list[str] = []
+        for id, item in items.items():
+            if not isinstance(item.description, str) or not item.description.strip():
+                item.description = "(空)"
+                repaired.append(f"条目 {id}: description 为空，已重置")
+
+            if not isinstance(item.theme, str) or not item.theme.strip():
+                item.theme = "(未分类)"
+                repaired.append(f"条目 {id}: theme 为空，已重置")
+
+            if not isinstance(item.history, list):
+                item.history = []
+                repaired.append(f"条目 {id}: history 非列表，已重置")
+
+            if (
+                not isinstance(item.latest_update_time, str)
+                or not item.latest_update_time.strip()
+            ):
+                item.latest_update_time = datetime.datetime.now().strftime(
+                    "%Y-%m-%d %H:%M:%S"
+                )
+                repaired.append(f"条目 {id}: latest_update_time 无效，已重置")
+
+            if not isinstance(item.hit, int) or item.hit < 0:
+                item.hit = 0
+                repaired.append(f"条目 {id}: hit 无效，已重置为 0")
+
+        return issues, repaired
 
     # ── 默认 ID 生成 ────────────────────────────────────────
 
@@ -182,17 +225,27 @@ class BaseMemoryManager(ABC):
         ]
         return {"sections": sections}
 
-    # ── 抽象 CRUD 方法（子类可 override 以优化） ─────────────
+    # ── CRUD 方法（基于 _load_all / _save_all / _write_lock） ──
 
-    @abstractmethod
     def add(self, description: str, theme: str) -> str:
-        ...
+        """添加一条新的记忆条目。"""
+        with self._write_lock():
+            items = self._load_all()
+            new_id = self._generate_id()
+            items[new_id] = MemoryItem(description, theme)
+            self._save_all(items)
+        return new_id
 
-    @abstractmethod
     def delete(self, id: str) -> str:
-        ...
+        """删除指定 ID 的记忆条目，返回被删除条目的描述。"""
+        with self._write_lock():
+            items = self._load_all()
+            if id not in items:
+                raise ValueError(f"Memory item with ID {id} not found")
+            removed = items.pop(id)
+            self._save_all(items)
+        return removed.description
 
-    @abstractmethod
     def merge(
         self,
         id1: str,
@@ -201,9 +254,17 @@ class BaseMemoryManager(ABC):
         merged_theme: str,
         reason: str,
     ) -> None:
-        ...
+        """将两条记忆合并为一条，id1 保留，id2 被删除。"""
+        with self._write_lock():
+            items = self._load_all()
+            if id1 not in items or id2 not in items:
+                raise ValueError(
+                    f"Memory items with IDs {id1} and {id2} not found"
+                )
+            items[id1].merge(items[id2], reason, merged_description, merged_theme)
+            items.pop(id2)
+            self._save_all(items)
 
-    @abstractmethod
     def update(
         self,
         id: str,
@@ -211,9 +272,20 @@ class BaseMemoryManager(ABC):
         new_description: str | None = None,
         new_theme: str | None = None,
     ) -> None:
-        ...
+        """更新指定记忆条目的内容和/或主题。"""
+        with self._write_lock():
+            items = self._load_all()
+            if id not in items:
+                raise ValueError(f"Memory item with ID {id} not found")
+            items[id].update(reason, new_description, new_theme)
+            self._save_all(items)
 
-    @abstractmethod
     def hit(self, id: str) -> int:
         """将指定记忆的 hit 计数加一，返回新的计数。"""
-        ...
+        with self._write_lock():
+            items = self._load_all()
+            if id not in items:
+                raise ValueError(f"Memory item with ID {id} not found")
+            items[id].hit += 1
+            self._save_all(items)
+            return items[id].hit
